+++ generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/api/v36/converters/ReportsConverter.java
@@ -23,7 +23,6 @@ public class ReportsConverter {
     private static final Map<Reports.IssuesFilter.SuppressStatus, IssuesFilterSuppressStatus> REVERSE_ISSUE_FILTER_SUPPRESSSTATUS_MAP = new HashMap<>();
     private static final Map<Reports.IssuesFilter.SourceType, IssuesFilterSourceType> REVERSE_ISSUE_FILTER_SOURCETYPE_MAP = new HashMap<>();
     private static final Map<Reports.Report.Format, ReportFormatType> REVERSE_REPORT_FORMAT_MAP = new HashMap<>();
-    private static final Map<Reports.Data.Format, ReportFormatType> REVERSE_DATA_FORMAT_MAP = new HashMap<>();
 
     static {
         REVERSE_ISSUE_FILTER_LEVEL_MAP.put(Reports.IssuesFilter.Level.NONE, IssuesFilterLevel.None);
@@ -66,20 +65,11 @@ public class ReportsConverter {
         REVERSE_ISSUE_FILTER_SOURCETYPE_MAP.put(Reports.IssuesFilter.SourceType.ALL, IssuesFilterSourceType.All);
 
         REVERSE_REPORT_FORMAT_MAP.put(Reports.Report.Format.HTML, ReportFormatType.HTML);
-        REVERSE_REPORT_FORMAT_MAP.put(Reports.Report.Format.PDF, ReportFormatType.PDF);
-
-        REVERSE_DATA_FORMAT_MAP.put(Reports.Data.Format.JSON, ReportFormatType.JSON);
-        REVERSE_DATA_FORMAT_MAP.put(Reports.Data.Format.XML, ReportFormatType.XML);
     }
 
     @NonNull
     public static ReportFormatType convert(@NonNull final Reports.Report.Format format) {
         return REVERSE_REPORT_FORMAT_MAP.get(format);
-    }
-
-    @NonNull
-    public static ReportFormatType convert(@NonNull final Reports.Data.Format format) {
-        return REVERSE_DATA_FORMAT_MAP.get(format);
     }
 
     @NonNull