+++ ptai-cli-plugin/src/test/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/GenerateReportIT.java
@@ -27,10 +27,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Stream;
 
-import static com.ptsecurity.appsec.ai.ee.scan.reports.Reports.Data.Format.JSON;
-import static com.ptsecurity.appsec.ai.ee.scan.reports.Reports.Data.Format.XML;
 import static com.ptsecurity.appsec.ai.ee.scan.reports.Reports.Locale.EN;
 import static com.ptsecurity.appsec.ai.ee.scan.reports.Reports.Locale.RU;
 import static com.ptsecurity.appsec.ai.ee.scan.reports.Reports.Report.Format.HTML;
@@ -65,7 +62,7 @@ class GenerateReportIT extends BaseCliIT {
     }
 
     @Test
-    @DisplayName("Generate specific app01 scan results report using all possible configurations")
+    @DisplayName("Generate specific  scan results report using all possible configurations")
     public void testSpecificScanResultsReport() {
         UUID scanResultId = getLatestCompleteScanResults(BaseAstIT.PHP_SMOKE_MEDIUM.getName());
         testScanResultsReport(scanResultId);
@@ -82,13 +79,6 @@ class GenerateReportIT extends BaseCliIT {
         pair.getRight().addAll(Arrays.asList("--raw-data-file", pair.getLeft()));
         cases.add(pair);
 
-        pair = new ImmutablePair<>("data-en.xml", new ArrayList<>());
-        pair.getRight().addAll(Arrays.asList(
-                "--data-file", pair.getLeft(),
-                "--data-locale", EN.name(),
-                "--data-format", XML.name()));
-        cases.add(pair);
-
         pair = new ImmutablePair<>("report-ru.html", new ArrayList<>());
         pair.getRight().addAll(Arrays.asList(
                 "--report-file", pair.getLeft(),
@@ -97,7 +87,7 @@ class GenerateReportIT extends BaseCliIT {
                 "--report-format", HTML.name()));
         cases.add(pair);
 
-        for (int i = 1 ; i < 7 ; i++) {
+        for (int i = 1 ; i < 3 ; i++) {
             Path folder = Paths.get(destination.toString()).resolve(UUID.randomUUID().toString());
             List<String> args = new ArrayList<>(Arrays.asList(
                     "generate-report",
@@ -110,12 +100,12 @@ class GenerateReportIT extends BaseCliIT {
                 args.add("--scan-result-id");
                 args.add(scanResultId.toString());
             }
-            for (int j = 0 ; j < 3 ; j++)
+            for (int j = 0 ; j < 2 ; j++)
                 if (0 != (i & 1 << j))
                     args.addAll(cases.get(j).getRight());
             Integer res = new CommandLine(new Plugin()).execute(args.toArray(new String[0]));
             Assertions.assertEquals(BaseCommand.ExitCode.SUCCESS.getCode(), res);
-            for (int j = 0 ; j < 3 ; j++)
+            for (int j = 0 ; j < 2 ; j++)
                 if (0 != (i & 1 << j))
                     Assertions.assertTrue(folder.resolve(cases.get(j).getLeft()).toFile().exists());
         }
@@ -145,9 +135,6 @@ class GenerateReportIT extends BaseCliIT {
                 "--report-file", "owasp.ru.html",
                 "--report-locale", EN.name(),
                 "--report-format", HTML.name(),
-                "--data-file", "owasp.ru.html",
-                "--data-locale", EN.name(),
-                "--data-format", JSON.name(),
                 "--raw-data-file", "raw.json");
         Assertions.assertEquals(BaseCommand.ExitCode.FAILED.getCode(), res);
     }
@@ -213,7 +200,7 @@ class GenerateReportIT extends BaseCliIT {
 
     @SneakyThrows
     @Test
-    @DisplayName("Generate JSON-defined reports for specific app01 scan results using extended filters")
+    @DisplayName("Generate JSON-defined reports for specific PHP smoke medium scan results using extended filters")
     public void testLatestExtendedJsonDefinedReportsGeneration(@NonNull final TestInfo testInfo) {
         log.trace(testInfo.getDisplayName());
         Path reportsJson = TEMP_FOLDER().resolve(UUID.randomUUID().toString());
@@ -359,7 +346,7 @@ class GenerateReportIT extends BaseCliIT {
     protected static void checkReports(@NonNull final Path reportsJson, @NonNull final Path destination) {
         ObjectMapper mapper = BaseJsonHelper.createObjectMapper();
         Reports reports = mapper.readValue(reportsJson.toFile(), Reports.class);
-        Stream.concat(reports.getReport().stream(), reports.getData().stream()).forEach((r) -> {
+        reports.getReport().forEach((r) -> {
             File report = destination.resolve(r.getFileName()).toFile();
             Assertions.assertTrue(report.exists());
         });