+++ ptai-data-structures/src/test/java/com/ptsecurity/appsec/ai/ee/test/scan/reports/ReportsTest.java
@@ -35,11 +35,9 @@ public class ReportsTest extends BaseTest {
 
         Assertions.assertNotNull(reports);
         Assertions.assertEquals(1, reports.getReport().size());
-        Assertions.assertEquals(2, reports.getData().size());
         Assertions.assertEquals(1, reports.getRaw().size());
 
         Assertions.assertEquals(Reports.Report.Format.HTML, reports.getReport().get(0).getFormat());
-        Assertions.assertEquals(Reports.Data.Format.XML, reports.getData().get(1).getFormat());
     }
 
     @Test